+++ src/main/java/com/oya/kr/community/domain/CommunityAdvertisement.java
@@ -1,9 +1,8 @@
-package com.oya.kr.commutiny.domain;
+package com.oya.kr.community.domain;
 
 import org.springframework.stereotype.Component;
 
 import com.oya.kr.global.domain.Base;
-import com.oya.kr.user.domain.User;
 
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -12,9 +11,9 @@ import lombok.NoArgsConstructor;
 @Getter
 @Component
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class CommunityView extends Base {
+public class CommunityAdvertisement extends Base {
 
 	private Long id;
 	private Community community;
-	private User user;
+	private int amount;
 }
