+++ src/main/java/com/oya/kr/community/domain/CommunityView.java
@@ -1,8 +1,9 @@
-package com.oya.kr.commutiny.domain;
+package com.oya.kr.community.domain;
 
 import org.springframework.stereotype.Component;
 
 import com.oya.kr.global.domain.Base;
+import com.oya.kr.user.domain.User;
 
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -11,9 +12,9 @@ import lombok.NoArgsConstructor;
 @Getter
 @Component
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class Vote extends Base {
+public class CommunityView extends Base {
 
 	private Long id;
 	private Community community;
-	private String content;
+	private User user;
 }
